requirements.txt
torch
wheel
numpy
pandas
scikit-learn
opencv-python
selenium
webdriver-manager
python-dotenv
Pillow
webp
torchvision